refactor(mitm): extract privileged step helpers to satisfy file-size cap
manager.ts exceeded the 800-line cap after #7938 gates. Move DNS teardown
and the shared sudo skip runner into dedicated modules; behavior unchanged.
