+++ src/mitm/manager.ts
@@ -22,7 +22,8 @@ import {
   performRepairSteps,
   type RepairPlan,
 } from "./repair.ts";
-import { canRunPrivilegedMitmSteps } from "./sudoGate.ts";
+import { runPrivilegedMitmStep } from "./privilegedMitmStep.ts";
+import { removeStopDnsEntries } from "./stopDnsTeardown.ts";
 
 export { buildRepairPlan, collectManagedHosts, type RepairPlan };
 
@@ -520,41 +521,43 @@ async function startMitmInternal(
   //    bridge: in containers/headless the system trust store can't be written,
   //    so we start in "untrusted" mode and let the operator trust the CA by hand
   //    (mirrors the best-effort "continuing" pattern used for DNS below). (#4546)
-  //    Skip entirely when no sudo password is available — never spawn sudo -S
-  //    with an empty string (#7938).
   let certTrusted = false;
-  if (canRunPrivilegedMitmSteps(sudoPassword)) {
-    try {
-      const certResult =
-        migrationDecision === "use-root-ca"
-          ? await installCaCert(sudoPassword, certPath)
-          : await installCertResult(sudoPassword, certPath);
-      certTrusted = certResult.installed;
-      if (!certResult.installed) {
-        log.warn(
-          { reason: certResult.reason },
-          "MITM cert not auto-trusted; bridge starting in skip mode (manual trust required)"
-        );
+  await runPrivilegedMitmStep(
+    sudoPassword,
+    "Skipping MITM cert trust — no sudo password available (#7938)",
+    async () => {
+      try {
+        const certResult =
+          migrationDecision === "use-root-ca"
+            ? await installCaCert(sudoPassword, certPath)
+            : await installCertResult(sudoPassword, certPath);
+        certTrusted = certResult.installed;
+        if (!certResult.installed) {
+          log.warn(
+            { reason: certResult.reason },
+            "MITM cert not auto-trusted; bridge starting in skip mode (manual trust required)"
+          );
+        }
+      } catch (err) {
+        log.error({ err }, "installCertResult threw unexpectedly (continuing without trusted cert)");
       }
-    } catch (err) {
-      log.error({ err }, "installCertResult threw unexpectedly (continuing without trusted cert)");
     }
-  } else {
-    log.info("Skipping MITM cert trust — no sudo password available (#7938)");
-  }
+  );
 
   // 3. Add DNS entries: Antigravity defaults + all agents with dns_enabled=true +
   //    all custom hosts with enabled=true. Best-effort — see provisionDnsEntries.
-  if (canRunPrivilegedMitmSteps(sudoPassword)) {
-    log.info("Adding DNS entries...");
-    try {
-      await provisionDnsEntries(sudoPassword);
-    } catch (err) {
-      log.error({ err }, "DNS provisioning threw unexpectedly (continuing)");
+  await runPrivilegedMitmStep(
+    sudoPassword,
+    "Skipping DNS provisioning — no sudo password available (#7938)",
+    async () => {
+      log.info("Adding DNS entries...");
+      try {
+        await provisionDnsEntries(sudoPassword);
+      } catch (err) {
+        log.error({ err }, "DNS provisioning threw unexpectedly (continuing)");
+      }
     }
-  } else {
-    log.info("Skipping DNS provisioning — no sudo password available (#7938)");
-  }
+  );
 
   // 4. Start MITM server
   log.info("Starting MITM server...");
@@ -682,31 +685,6 @@ async function startMitmInternal(
 }
 
 /**
- * DNS teardown step of stopMitm() (#1809) — split out purely to keep
- * stopMitm()'s own cyclomatic complexity under the repo's ratchet; behavior
- * is unchanged from the original inline implementation.
- */
-async function removeStopDnsEntries(
-  deps: {
-    removeDNSEntry: (sudoPassword: string) => Promise<void>;
-    removeDNSEntries: (hosts: string[], sudoPassword: string) => Promise<void>;
-    collectManagedHosts: () => string[];
-  },
-  sudoPassword: string
-): Promise<void> {
-  log.info("Removing DNS entries...");
-  await deps.removeDNSEntry(sudoPassword);
-  try {
-    const managed = deps.collectManagedHosts();
-    if (managed.length > 0) {
-      await deps.removeDNSEntries(managed, sudoPassword);
-    }
-  } catch (err) {
-    log.error({ err }, "Failed to remove managed DNS entries during stop (continuing)");
-  }
-}
-
-/**
  * Kill the MITM server process during stop — either the in-memory
  * `serverProcess` handle or, if that's gone, the PID recorded in `PID_FILE`.
  * Split out of stopMitm() purely to keep that function's complexity under
@@ -776,16 +754,12 @@ export async function stopMitm(
     collectManagedHosts: _depsOverride?.collectManagedHosts ?? collectManagedHosts,
   };
 
-  // 1. Remove DNS entries FIRST — see function doc + module doc above for why
-  //    this must happen before the process kill (#1809, Gap 8). Skip when no
-  //    sudo password is available — never spawn sudo -S with an empty string
-  //    (#7938). Stale /etc/hosts entries may remain until stop with a password
-  //    or Repair; the server kill below still runs so the listener is torn down.
-  if (canRunPrivilegedMitmSteps(sudoPassword)) {
-    await removeStopDnsEntries(deps, sudoPassword);
-  } else {
-    log.info("Skipping DNS teardown — no sudo password available (#7938)");
-  }
+  // 1. Remove DNS entries FIRST — see function doc + module doc above (#1809).
+  await runPrivilegedMitmStep(
+    sudoPassword,
+    "Skipping DNS teardown — no sudo password available (#7938)",
+    () => removeStopDnsEntries(deps, sudoPassword)
+  );
 
   // 2. Kill server process (in-memory or from PID file)
   await killMitmServerProcessOnStop();
